refactor(scanner)!: Move `ScanResultsStorage` to storage package
Move the abstract base class for package based scan storages to the
storage package, next to the classes that inherit from it.
